Reduce samples dependencies on everything
